connected ajax GET && POST
I connected the GET portion of the API where we are using ajax to communicate and the list is retrieved and added to the public chat box upon load, refresh, and 5 second timed refresh. additionally the user is able to create public chats and those chats are updated visually every 5 second or upon hitting refresh.